refactor(types): rename and relocate types
Rename `Transfer` to `TransferHistoryItem` for clarity.

Move interface from types.ts:
 - `TransferFormProps` interface to TransferForm.tsx,
 - `TransferPage` interface to TransferPageComponent.tsx file.
